+++ src/cmd/init.rs
@@ -4,11 +4,13 @@ use crate::{
     cache::CacheMode,
     utils, Cache, Caching, Client, Config, Fetch, InitArgs, SourceBookmarks, TargetBookmarks,
 };
-use log::info;
+use log::debug;
 
 /// Import bookmarks, fetch bookmarks from url, and save fetched websites in
 /// cache if bookmarks were not imported yet.
 pub async fn init(config: &Config, args: &InitArgs) -> Result<(), anyhow::Error> {
+    debug!("{args:?}");
+
     let mut source_reader = config
         .settings
         .sources
@@ -22,7 +24,7 @@ pub async fn init(config: &Config, args: &InitArgs) -> Result<(), anyhow::Error>
     target_reader.read(&mut target_bookmarks)?;
 
     if !target_bookmarks.is_empty() {
-        info!("Bookmarks already imported");
+        println!("Bookmarks already imported");
     } else {
         let cache_mode = CacheMode::new(&args.mode, &config.settings.cache_mode);
         let cache = Cache::new(&config.cache_path, cache_mode);
@@ -59,7 +61,7 @@ async fn init_bookmarks(
 
     let mut target_bookmarks = TargetBookmarks::from(source_bookmarks);
 
-    info!(
+    println!(
         "Imported {} bookmarks from {} sources: {}",
         target_bookmarks.len(),
         source_reader.len(),