+++ src/lib.rs
@@ -32,10 +32,15 @@ mod client;
 pub mod cmd;
 /// The configuration used in Bogrep.
 mod config;
+/// The errors which can occur in Bogrep.
+pub mod errors;
 /// Helper functions to work with HTML.
 pub mod html;
 /// Helper function to work with JSON.
 pub mod json;
+/// Initialize a simple logger based on the verbosity level (or the `RUST_LOG`
+/// environment variable).
+mod logger;
 /// The settings used in Bogrep.
 mod settings;
 /// Utilities used in testing.
@@ -54,4 +59,5 @@ pub use bookmarks::{
 pub use cache::{Cache, Caching, MockCache};
 pub use client::{Client, Fetch, MockClient};
 pub use config::Config;
+pub use logger::Logger;
 pub use settings::Settings;